+++ src/config.js
@@ -2,7 +2,8 @@
 
 let config = {
   userName: 'Debug User',
-  keyBindings: { /* TODO - Add some Key Bindings, so I can use this, rather than hard-coded inputs */ }
+  keyBindings: { /* TODO - Add some Key Bindings, so I can use this, rather than hard-coded inputs */ },
+  screenSize: 4 // How much larger the screen is than the actual
 }
 
 export default config;